backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
scikit-learn==1.3.2
xgboost==2.0.3
pandas>=2.0.0
numpy>=1.20.0
joblib==1.3.2
aiosmtplib==3.0.1
email-validator==2.1.0
